create tools.py, move some functions there
